refactor CreateEvent to utilize EventForm component, commit is pre test migration
